Move removeSoftDeletedBALs to lib/models/base-locale
