+++ cron.js
@@ -2,7 +2,7 @@
 require('dotenv').config()
 const ms = require('ms')
 const {detectOutdated, detectConflict, syncOutdated} = require('./lib/sync')
-const {removeSoftDeletedBALs} = require('./lib/util/remove-soft-deleted-bals.js')
+const {removeSoftDeletedBALs} = require('./lib/models/base-locale')
 const mongo = require('./lib/util/mongo')
 
 const jobs = [
