+++ lib/models/base-locale.js
@@ -1,4 +1,5 @@
 const Joi = require('joi')
+const subMonths = require('date-fns/subMonths')
 const {omit, difference, uniq, keyBy} = require('lodash')
 const {generateBase62String} = require('../util/base62')
 const {validPayload} = require('../util/payload')
@@ -464,6 +465,17 @@ async function updateHabilitation(baseLocaleId, habilitationId) {
   )
 }
 
+async function removeSoftDeletedBALs() {
+  const now = new Date()
+  const deleteTime = subMonths(now, 12)
+
+  const softDeletedBALs = await mongo.db.collection('bases_locales').deleteMany({
+    _deleted: {$lt: deleteTime}
+  })
+
+  console.log(`\n${softDeletedBALs.deletedCount} BALs supprimées définitivement\n`)
+}
+
 module.exports = {
   create,
   createSchema,
@@ -493,5 +505,6 @@ module.exports = {
   getNumerosCount,
   getExpandedBAL,
   batchUpdateNumeros,
-  updateHabilitation
+  updateHabilitation,
+  removeSoftDeletedBALs
 }
